Create pack crash fix

+++ ReskinMaker/ReskinMaker/CreateSkinPackForm.cs
@@ -88,19 +88,19 @@ namespace ReskinMaker
         //haHAA
         const string modcs = "using System;\nusing System.Collections.Generic;\nusing System.Linq;\nusing System.Text;\nusing System.Reflection;\nusing System.IO;\nusing System.Windows.Forms;\nusing System.Runtime.InteropServices;\nusing System.Text.RegularExpressions;\nusing System.ComponentModel;\nusing System.Diagnostics;\nusing Microsoft.Xna.Framework;\n[assembly: AssemblyTitle(\"@NAME@\")]\n[assembly: AssemblyCompany(\"@CREATOR@\")]\n[assembly: AssemblyDescription(\"@DESCRIPTION@\")]\n[assembly: AssemblyVersion(\"@VERSION@\")]\nnamespace DuckGame.MyMod{\npublic class MyMod : DisabledMod, IUpdateable{\npublic static ModConfiguration config;\nstatic PropertyInfo steamIdField = typeof(ModConfiguration).GetProperty(\"workshopID\",BindingFlags.Instance|BindingFlags.NonPublic);\nstatic PropertyInfo disabledField = typeof(ModConfiguration).GetProperty(\"disabled\", BindingFlags.Instance|BindingFlags.NonPublic);\npublic static string replaceData{get{\nreturn config.isWorkshop ? steamIdField.GetValue(config, new object[0]).ToString() : \"LOCAL\";}}\npublic static bool disabled{\nset { disabledField.SetValue(config,value,new object[0]); }\nget { return (bool)disabledField.GetValue(config, new object[0]); }}\nconst BindingFlags flags = BindingFlags.Static | BindingFlags.NonPublic | BindingFlags.Instance;\nprotected override void OnPostInitialize(){\nconfig = configuration;\nvar form = (Form)Control.FromHandle(MonoMain.instance.Window.Handle);\nform.FormClosing += FormClosed;\n(typeof(Game).GetField(\"updateableComponents\", flags).GetValue(MonoMain.instance) as List<IUpdateable>).Add(this);}\nvoid FormClosed(object sender,EventArgs e){\nif (!Program.commandLine.Contains(\"-download\")) return;disabled = false;\ntypeof(ModLoader).GetMethod(\"DisabledModsChanged\",flags).Invoke(null,new object[0]);}\npublic bool Enabled{get{return true;}}\npublic int UpdateOrder{get{\nreturn 1;}}\npublic event EventHandler<EventArgs> EnabledChanged;\npublic event EventHandler<EventArgs> UpdateOrderChanged;bool updateLobby;\npublic void Update(GameTime gameTime){\nif (ModLoader.modsEnabled && Level.current is TeamSelect2 && Steam.lobby != null && Steam.lobby.id != 0L){\nstring str;int n;\nif (updateLobby && !string.IsNullOrEmpty((str = Steam.lobby.GetLobbyData(\"mods\"))) && (n = str.IndexOf(replaceData)) > -1){\nstr = str.Remove(n, replaceData.Length).Trim('|').Replace(\"||\", \"|\");\nSteam.lobby.SetLobbyData(\"mods\", str);\nupdateLobby = false;}}\nelse updateLobby = true;}}}";
         const string modconf = "<Mod>\n<HardDependencies>reskins</HardDependencies>\n<Mod>";
-        void Create(string name,string creator,string description,string version)
+        bool Create(string name,string creator,string description,string version)
         {
-            if(name.IndexOfAny(Path.GetInvalidPathChars()) >= 0)
+            if(name.IndexOfAny(Path.GetInvalidPathChars()) >= 0 || name.IndexOfAny(Path.GetInvalidFileNameChars()) >= 0)
             {
-                ShowWarning("Name cannot contain any of these characters: \n"+String.Join(" ",Path.GetInvalidPathChars()),"Invalid Name");
-                return;
+                ShowWarning("Name cannot contain any of these characters: \n"+String.Join(" ",Path.GetInvalidPathChars()) +" "+ String.Join(" ",Path.GetInvalidFileNameChars()), "Invalid Name");
+                return false;
             }
 
             string basePath = Environment.GetFolderPath(Environment.SpecialFolder.Personal)+@"\DuckGame\Mods\";
             string path = basePath + name;
             if (Directory.Exists(path))
             {
-                if (MessageBox.Show("This pack already exists\nDo you want to overwrite?", "warning!", MessageBoxButtons.YesNo, MessageBoxIcon.Warning) == DialogResult.No) return;
+                if (MessageBox.Show("This pack already exists\nDo you want to overwrite?", "warning!", MessageBoxButtons.YesNo, MessageBoxIcon.Warning) == DialogResult.No) return false;
                 Directory.Delete(path,true);
             }
 
@@ -111,6 +111,7 @@ namespace ReskinMaker
 
             File.WriteAllText(path+@"\build\src\Mod.cs",modcs.Replace("@NAME@",name).Replace("@CREATOR@",creator.Replace("\"","\\\"")).Replace("@DESCRIPTION@",description.Replace("\"", "\\\"")).Replace("@VERSION@",version.Replace("\"", "\\\"")));
             MessageBox.Show("Successfully created pack!\nOpen duckgame and it will appear in the mod list");
+            return true;
         }
 
         void CopyDir(string SourcePath,string DestinationPath)
@@ -131,7 +132,7 @@ namespace ReskinMaker
                 return;
             }
 
-            Create(TextName.Text,TextAuthor.Text,TextDescription.Text,TextVersion.Text);
+            if (!Create(TextName.Text, TextAuthor.Text, TextDescription.Text, TextVersion.Text)) return;
 
             DialogResult = DialogResult.OK;
             this.Close();
